Add files via upload
This myImageSVD.py is a little bit altered from an original program provided in https://github.com/compsci-alliance/information-compression/ .  The original is an example code for a mid-term report of a class named "information compression" held in Univ. Tokyo for graduate students of "Science" and "Information Science and Technology".  I left my code here just for record. If there is a problem, I'll delete it.